requirements.txt
streamlit>=1.30.0
python-docx>=1.1.0
anthropic>=0.40.0
requests>=2.31.0
Pillow>=10.0.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
markdown>=3.5.0
